# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TOPIDEALSPACE API",
        "description": "API REST de la marketplace TOPIDEALSPACE. Authentification via Laravel Sanctum (Bearer Token). Connectez-vous via **POST /auth/login**, copiez le token retourné, puis cliquez sur **Authorize** en haut à droite.",
        "contact": {
            "email": "contact@topidealspace.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur local (dev)"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Produits Publics"
                ],
                "summary": "Liste des produits publiés",
                "description": "Récupère la liste paginée des produits validés et visibles publiquement",
                "operationId": "getPublicProducts",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre de produits par page",
                        "schema": {
                            "type": "integer",
                            "default": 12
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par nom ou description",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Critère de tri",
                        "schema": {
                            "type": "string",
                            "default": "recent",
                            "enum": [
                                "recent",
                                "price_asc",
                                "price_desc",
                                "name"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des produits",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Produit"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Produits Publics"
                ],
                "summary": "Détail d'un produit",
                "description": "Récupère les détails complets d'un produit publié",
                "operationId": "getPublicProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du produit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Produit"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produit non trouvé ou non publié"
                    }
                }
            }
        },
        "/products/{id}/view": {
            "post": {
                "tags": [
                    "Produits Publics"
                ],
                "summary": "Enregistrer une vue",
                "description": "Enregistre une vue sur un produit. Limité à 1 vue par IP par produit toutes les 30 minutes.",
                "operationId": "recordProductView",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vue enregistrée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vue enregistrée avec succès"
                                        },
                                        "vues_count": {
                                            "type": "integer",
                                            "example": 42
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    },
                    "429": {
                        "description": "Vue déjà enregistrée (rate limit)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vue déjà enregistrée"
                                        },
                                        "vues_count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/suggestions": {
            "get": {
                "tags": [
                    "Produits Publics"
                ],
                "summary": "Suggestions de recherche",
                "description": "Retourne jusqu'à 10 produits correspondant à la recherche (minimum 2 caractères)",
                "operationId": "getProductSuggestions",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Terme de recherche (min 2 caractères)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des suggestions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_produits": {
                                                        "type": "integer"
                                                    },
                                                    "nom_produits": {
                                                        "type": "string"
                                                    },
                                                    "prix_produits": {
                                                        "type": "number"
                                                    },
                                                    "image_produits": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Produit": {
                "properties": {
                    "id_produits": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom_produits": {
                        "type": "string",
                        "example": "iPhone 15 Pro"
                    },
                    "prix_produits": {
                        "type": "number",
                        "format": "float",
                        "example": 599000
                    },
                    "description_produits": {
                        "type": "string",
                        "example": "Smartphone haut de gamme..."
                    },
                    "image_produits": {
                        "type": "string",
                        "example": "https://cloudinary.com/iphone.jpg"
                    },
                    "image_produits1": {
                        "type": "string",
                        "nullable": true
                    },
                    "image_produits2": {
                        "type": "string",
                        "nullable": true
                    },
                    "id_sous_categorie": {
                        "type": "integer",
                        "example": 5
                    },
                    "id_client": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_user": {
                        "type": "integer",
                        "nullable": true
                    },
                    "id_state": {
                        "type": "integer",
                        "example": 1
                    },
                    "dateSaisie": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentification via Laravel Sanctum Bearer Token. Entrez le token obtenu lors de la connexion.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}